Moved layer 8 init function. Removed some more unused code
